+++ src/Mew/Logic/Preset.ts
@@ -1,0 +1,17 @@
+import {BlockElement} from "./BlockElement";
+
+export class Preset {
+    tag: string
+    output: BlockElement
+    callback: CallableFunction
+
+    constructor(
+        tag: string,
+        output: BlockElement = new BlockElement(),
+        callback: CallableFunction = (r, old) => r
+    ) {
+        this.tag = tag
+        this.output = output
+        this.callback = callback
+    }
+}